revert unnecessary account refactor
