+++ account/priv_account.go
@@ -14,6 +14,10 @@
 
 package account
 
+// TODO: [ben] Account and PrivateAccount need to become a pure interface
+// and then move the implementation to the manager types.
+// Eg, Geth has its accounts, different from BurrowMint
+
 import (
 	"fmt"
 
@@ -23,6 +27,12 @@ import (
 	"github.com/tendermint/go-crypto"
 	"github.com/tendermint/go-wire"
 )
+
+type PrivAccount struct {
+	Address []byte         `json:"address"`
+	PubKey  crypto.PubKey  `json:"pub_key"`
+	PrivKey crypto.PrivKey `json:"priv_key"`
+}
 
 func (pA *PrivAccount) Generate(index int) *PrivAccount {
 	newPrivKey := pA.PrivKey.(crypto.PrivKeyEd25519).Generate(index)
